requirements.txt
aiofiles==0.6.0
aioredis==1.3.1
aiosmtplib==1.1.4
async-timeout==3.0.1
blinker==1.4
certifi==2020.12.5
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fakeredis==1.4.5
fastapi==0.63.0
fastapi-mail==0.3.4.2
h11==0.12.0
hiredis==1.1.0
httpcore==0.12.3
httpx==0.16.1
idna==3.1
Jinja2==2.11.3
MarkupSafe==1.1.1
pydantic==1.7.3
python-decouple==3.4
python-multipart==0.0.5
redis==3.5.3
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.3.0
starlette==0.13.6
uvicorn==0.13.3
